requirements-prod.txt
aio-pika==9.5.6
alembic==1.16.5
async-lru==2.0.5
asyncpg==0.30.0
dramatiq[redis, watch]==1.18.0
fastapi-camelcase==2.0.0
fastapi[all]==0.118.0
gunicorn==23.0.0
periodiq==0.13.0
pydantic-settings==2.11.0
redis[hiredis]==6.4.0
safe-eth-py==7.11.0
setuptools>=75.0.0
sqladmin[full]==0.21.0
sqlmodel==0.0.25
uvicorn-worker==0.4.0
web3==7.13.0
